RF: Move test to match code location
