+++ datalad/distribution/tests/test_utils.py
@@ -13,7 +13,6 @@ import os
 from os.path import join as opj
 
 from datalad.distribution.utils import _get_flexible_source_candidates
-from datalad.distribution.utils import _get_flexible_source_candidates_for_submodule
 from datalad.distribution.utils import get_git_dir
 
 from datalad.tests.utils import with_tempfile
@@ -57,36 +56,6 @@ def test_get_flexible_source_candidates():
     # when source is not relative, but base_url is specified as just the destination path,
     # not really a "base url" as name was suggesting, then it should be ignored
     eq_(f('http://e.c/p', '/path'), ['http://e.c/p', 'http://e.c/p/.git'])
-
-
-@with_tempfile
-@with_tempfile
-def test_get_flexible_source_candidates_for_submodule(t, t2):
-    f = _get_flexible_source_candidates_for_submodule
-    # for now without mocking -- let's just really build a dataset
-    from datalad.api import create
-    from datalad.api import install
-    ds = create(t)
-    clone = install(t2, source=t)
-
-    # first one could just know about itself or explicit url provided
-    sshurl = 'ssh://e.c'
-    httpurl = 'http://e.c'
-    sm_httpurls = [httpurl, httpurl + '/.git']
-    eq_(f(ds, 'sub'), [])
-    eq_(f(ds, 'sub', sshurl), [sshurl])
-    eq_(f(ds, 'sub', httpurl), sm_httpurls)
-    eq_(f(ds, 'sub', None), [])  # otherwise really we have no clue were to get from
-
-    # but if we work on dsclone then it should also add urls deduced from its
-    # own location default remote for current branch
-    eq_(f(clone, 'sub'), [t + '/sub'])
-    eq_(f(clone, 'sub', sshurl), [t + '/sub', sshurl])
-    eq_(f(clone, 'sub', httpurl), [t + '/sub'] + sm_httpurls)
-    eq_(f(clone, 'sub'), [t + '/sub'])  # otherwise really we have no clue were to get from
-    # TODO: check that http:// urls for the dataset itself get resolved
-
-    # TODO: many more!!
 
 
 @with_tempfile
